+++ src/main/java/homework/mstruebing/app/domain/model/Password.java
@@ -6,7 +6,7 @@ package homework.mstruebing.app;
  */
 public class Password 
 {
-	protected long id;
+	protected int id;
 	protected PasswordList passwordList;
 	protected String title;
 	protected String username;
